requirements.txt
pandas==2.2.2
numpy==1.26.4
supabase==2.7.4
python-dotenv==1.0.1
statsmodels==0.14.2